pyproject.toml
[project]
name = "backup-orchestrator"
version = "0.1.0"
description = "Centralized backup orchestrator for go_backup agents over MQTT"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "paho-mqtt>=2.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
